requirements.txt
click==8.1.3
configparser==5.2.0
fake-useragent==1.1.1
pycryptodome==3.17
requests==2.28.1